+++ app/src/main/java/com/android/quizcafe/feature/login/LoginRoute.kt
@@ -1,0 +1,48 @@
+package com.android.quizcafe.feature.login
+
+import android.widget.Toast
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.platform.LocalContext
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.android.quizcafe.R
+
+@Composable
+fun LoginRoute(
+    navigateToSignUp: () -> Unit,
+    navigateToHome: () -> Unit,
+    viewModel: LoginViewModel = hiltViewModel()
+) {
+    val context = LocalContext.current
+    val state by viewModel.state.collectAsState()
+
+    LaunchedEffect(Unit) {
+        viewModel.effect.collect { effect ->
+            when (effect) {
+                LoginEffect.NavigateToHome -> {
+                    navigateToHome()
+                    Toast.makeText(
+                        context, context.getString(R.string.success_login), Toast.LENGTH_SHORT
+                    ).show()
+                }
+
+                LoginEffect.NavigateToSignUp -> {
+                    navigateToSignUp()
+                }
+
+                is LoginEffect.ShowErrorDialog -> {
+                    Toast.makeText(
+                        context, context.getString(R.string.error_message), Toast.LENGTH_SHORT
+                    ).show()
+                }
+            }
+        }
+    }
+
+    LoginScreen(
+        state = state,
+        onIntent = viewModel::onIntent
+    )
+}
