+++ app/src/main/java/com/android/quizcafe/feature/signup/SignUpScreen.kt
@@ -1,8 +1,12 @@
 package com.android.quizcafe.feature.signup
 
-import android.widget.Toast
 import androidx.compose.animation.AnimatedContent
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
@@ -10,65 +14,25 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
 import com.android.quizcafe.R
+import com.android.quizcafe.core.designsystem.QuizCafeButton
+import com.android.quizcafe.core.designsystem.theme.QuizCafeTheme
 import com.android.quizcafe.core.ui.QuizCafeTopAppBar
 import com.android.quizcafe.core.ui.TopAppBarTitle
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun SignUpScreen(
-    viewModel: SignUpViewModel = hiltViewModel(),
-    navigateToLogin: () -> Unit,
+    step: Int,
+    state: SignUpViewState,
+    onIntent: (SignUpIntent) -> Unit
 ) {
-    val context = LocalContext.current
-    val state by viewModel.state.collectAsState()
-    var step by remember { mutableIntStateOf(0) }
-
-    LaunchedEffect(Unit) {
-        viewModel.effect.collect { effect ->
-            when (effect) {
-                is SignUpEffect.NavigateToPasswordInput -> {
-                    step = 1
-                    Toast.makeText(
-                        context,
-                        context.getString(R.string.success_email_verification),
-                        Toast.LENGTH_SHORT
-                    ).show()
-                }
-
-                is SignUpEffect.ShowErrorDialog -> {
-                    Toast.makeText(
-                        context,
-                        context.getString(R.string.error_message),
-                        Toast.LENGTH_SHORT
-                    ).show()
-                }
-
-                is SignUpEffect.NavigateToLoginScreen -> {
-                    navigateToLogin()
-                    Toast.makeText(
-                        context,
-                        context.getString(R.string.success_signup),
-                        Toast.LENGTH_SHORT
-                    ).show()
-                }
-            }
-        }
-    }
-
     Scaffold(
         modifier = Modifier
             .fillMaxSize()
@@ -91,9 +55,9 @@ fun SignUpScreen(
                 state = state,
                 onClick = {
                     when {
-                        state.isSuccessVerification -> viewModel.onIntent(SignUpIntent.ClickSignUp)
-                        state.isCodeSent -> viewModel.onIntent(SignUpIntent.ClickVerifyCode)
-                        else -> viewModel.onIntent(SignUpIntent.ClickSendCode)
+                        state.isSuccessVerification -> onIntent(SignUpIntent.ClickSignUp)
+                        state.isCodeSent -> onIntent(SignUpIntent.ClickVerifyCode)
+                        else -> onIntent(SignUpIntent.ClickSendCode)
                     }
                 }
             )
@@ -101,15 +65,46 @@ fun SignUpScreen(
     ) { innerPadding ->
         AnimatedContent(step) {
             when (it) {
-                0 -> EmailInputContent(viewModel, state, innerPadding)
-                1 -> PasswordInputContent(viewModel, state, innerPadding)
+                0 -> EmailInputContent(state, onIntent, innerPadding)
+                1 -> PasswordInputContent(state, onIntent, innerPadding)
             }
+        }
+    }
+}
+
+@Composable
+fun BottomActionButton(state: SignUpViewState, onClick: () -> Unit) {
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .imePadding()
+            .navigationBarsPadding()
+    ) {
+        QuizCafeButton(
+            onClick = onClick,
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(56.dp),
+            enabled = if (state.isSuccessVerification) state.isSignUpEnabled else state.isNextEnabled
+        ) {
+            Text(stringResource(R.string.next))
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun SignUpScreenPreview() {
-    SignUpScreen(navigateToLogin = { })
+fun SignUpEmailVerificationPreview() {
+    QuizCafeTheme {
+        SignUpScreen(0, SignUpViewState()) {}
+    }
+
+}
+
+@Preview(showBackground = true)
+@Composable
+fun SignUpPasswordPreview() {
+    QuizCafeTheme {
+        SignUpScreen(1, SignUpViewState()) {}
+    }
 }
